requirements.txt
pyinfra==2.9.2
hydra-core==1.3.2
python-dotenv==1.0.1
rich==13.5.2